Add 6 trading improvements: fractional sizing, score-proportional positions, walk-forward backtest, max 3 positions, liquidity filter, no signal-exit on losers
- models.py: quantity columns changed Float (was Integer) for fractional crypto
- services/portfolio.py: qty type hints updated to float
- services/risk_manager.py: calculate_quantity returns fractional float (min 0.0001), score-proportional sizing (1%–4%), added is_high_volatility()
- config.py: POSITION_SIZE_MIN_PCT, POSITION_SIZE_MAX_PCT, MAX_OPEN_POSITIONS added
- services/trading.py: max positions gate, ATR volatility gate, liquidity hour filter (01-05 UTC weekends), signal-exit only on profitable positions, 50-EMA uptrend filter
- backtest.py: walk-forward OOS simulation on last 30% of bars, score-proportional qty, signal-exit fix, 50-EMA market regime gate

+++ models.py
@@ -18,7 +18,7 @@ class Trade(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     symbol = Column(String(50), nullable=False)
     type = Column(SQLEnum(TradeType), nullable=False)
-    quantity = Column(Integer, nullable=False)
+    quantity = Column(Float, nullable=False)
     price = Column(Float, nullable=False)
     total_value = Column(Float, nullable=False)
     realized_pnl = Column(Float, default=0.0)
@@ -29,7 +29,7 @@ class OpenPosition(Base):
     __tablename__ = "open_positions"
     id = Column(Integer, primary_key=True, autoincrement=True)
     symbol = Column(String(50), nullable=False)
-    quantity = Column(Integer, nullable=False)
+    quantity = Column(Float, nullable=False)
     entry_price = Column(Float, nullable=False)
     stop_loss_price = Column(Float, nullable=False)
     target_price = Column(Float, nullable=False)
